:truck: Move: getCommentList, getReflycommentList 이동
Move: getCommentList, getReflycommentList 이동
