+++ app/_apis/projectApi.ts
@@ -35,12 +35,12 @@ export const addProjectApi = {
 };
 
 export const editProjectApi = {
-  putProject: async (projectId: number, projectData: FormData) => {
-    const HEADER = getHeaders();
-    return await httpClient().putFormData(`/projects/${projectId}`, projectData, HEADER.headers);
-  },
   getProject: async (projectId: number) => {
     const HEADER = getHeaders();
     return await httpClient().get<EditProjectResponse>(`/projects/${projectId}/edits`, {}, HEADER.headers);
+  },
+  putProject: async (projectId: number, projectData: FormData) => {
+    const HEADER = getHeaders();
+    return await httpClient().putFormData(`/projects/${projectId}`, projectData, HEADER.headers);
   },
 };
